subtitle_inserter/requirements.txt
PySide6>=6.6
pandas>=1.5
chardet>=5.2
pysubs2>=1.6
Pillow>=10.0 